feat(lexer): 128-bit integer literals (B-2026-08-19-8 stage 2)
Stage 2 of implementing 128-bit integers. `ExprKind::Integer` and
`Token::Integer` carry i128, `Token::IntegerOutOfRange` carries u128, and a
magnitude past `u64::MAX` now lexes when it carries an explicit 128-bit suffix
-- in decimal, hex, binary and octal.

STILL NOT USABLE, BY DESIGN. `i128`/`u128` remain rejected at type-check
(B-2026-08-19-6), so a program containing `170141183460469231731687303715884105727i128`
lexes, parses, and is then refused on the type. The type must not become
nameable before it works; stage 5 lifts the rejection once codegen and the
runtime can carry the value.

GATED ON THE SUFFIX, and that gate is the whole reason this is safe. The
`(i64::MAX, u64::MAX]` band must keep producing `IntegerOutOfRange`, because the
parser encodes that band as a WRAPPED bit pattern on the signed carrier and the
typechecker reads it back through the suffix. Widening the lexer's thresholds
unconditionally would reroute every `u64` literal into a different token kind and
silently change that encoding -- a test now pins the band, since that is exactly
the simplification a later reader would reach for. An unsuffixed huge literal
still errors, as before.

THE WRAP SURVIVES ON PURPOSE. With an i128 node a `u64` magnitude would fit
positively and the wrap looks obsolete, but the INTERPRETER still represents
unsigned runtime values the same wrapped way. Retiring one without the other
would split the encoding between literal and value, so both move together at
stage 5; `literal_as_i128` says so where the read-back happens.

`narrow_literal_to_i64` is the AST-layer sibling of stage 1's `narrow_to_i64`:
~30 sites where a literal leaves the i128 node for an i64-only consumer, named
rather than cast so stage 5 has a greppable worklist, and a hard check so a
truncation cannot pass silently.

BOTH FEATURE LEGS, which is where this nearly went wrong: the default leg was
clean while `--features llvm` had 32 errors, because codegen consumes
`ExprKind::Integer` and is not compiled in the default build. CLAUDE.md warns
about exactly this and it still needed the second run to surface.

+++ src/query_attributes.rs
@@ -258,7 +258,7 @@ fn arg_to_record(a: &AttrArg) -> AttributeQueryArg {
 fn classify_value(kind: &ExprKind) -> AttributeQueryValue {
     match kind {
         ExprKind::StringLit(s) => AttributeQueryValue::String(s.clone()),
-        ExprKind::Integer(n, _) => AttributeQueryValue::Int(*n as i128),
+        ExprKind::Integer(n, _) => AttributeQueryValue::Int(*n),
         ExprKind::Bool(b) => AttributeQueryValue::Bool(*b),
         ExprKind::Path { segments, .. } => AttributeQueryValue::Path(segments.join(".")),
         ExprKind::Identifier(name) => AttributeQueryValue::Path(name.clone()),
